added instruction about server implementation
